Seperated the Hit.hh file into it's own source file
